requirements.txt
fastapi
unicorn
pydantic
request
python-dotenv
huggingface-hub 